Broke GameState out into its own file.
