+++ src/TicTacToe.Core/TicTacToe.Core.csproj
@@ -53,6 +53,7 @@
     <Compile Include="GameAction.cs" />
     <Compile Include="GameBoard.cs" />
     <Compile Include="Game.cs" />
+    <Compile Include="GameState.cs" />
     <Compile Include="HumanPlayer.cs" />
     <Compile Include="IPlayer.cs" />
     <Compile Include="Actions\OccupyGameAction.cs" />
